+++ CHANGELOG.md
@@ -27,7 +27,8 @@
 * Authentication
   * Updated the Microsoft Authentication Library (MSAL) dependency to version 1.3.0
 * Invoice
-  * Added the billing start and end dates to the invoice model  (Thanks [@msize](https://github.com/msize))
+  * Added the billing start and end dates to the invoice model (Thanks [@msize](https://github.com/msize))
+  * Added the ReservationOrderId property to the one time invoice line item model
 * Subscription
   * Added the missing get subscription usage record by resource configuration (Thanks [@msize](https://github.com/msize))